fix: reviewer's findings
